[fix]: 배포 실패 원인 찾기
